Time: 3 ms (74.78%), Space: 57.6 MB (65.58%) - LeetHub

+++ 0238-product-of-array-except-self/0238-product-of-array-except-self.java
@@ -14,7 +14,7 @@ class Solution {
         
         int left = 1;
         for(int i = 0; i<ans.length; ++i){
-            ans[i] = (int)(left * ans[i]);
+            ans[i] = (left * ans[i]);
             left *= nums[i];
         }
         return ans;
